added delete portfolio and show observation and upgrade security

+++ src/services/DeleteObservationService.js
@@ -3,7 +3,7 @@ const Portfolio = require('../models/Portfolio');
 const Observation = require('../models/Observation');
 
 class DeleteObservationService {
-  async execute({ observation_id, portfolio_id }) {
+  async execute({ observation_id, portfolio_id, educator_id }) {
     const portfolio = await Portfolio.findById(portfolio_id).populate(
       'educator',
       '-password',
@@ -11,6 +11,10 @@ class DeleteObservationService {
 
     if (!portfolio) {
       throw new AppError('Erro ao deletar Observação');
+    }
+
+    if (educator_id !== portfolio.educator.id) {
+      throw new AppError('Você não tem permissão para esta ação');
     }
 
     const observation = await Observation.findByIdAndDelete(observation_id);
